Move password storage to instance variable
We should be able to easily override the password storage location, both
as a user and during testing. For this reason I'm moving it to an
instance variable, which also required some reworking of the code so
that tests could still run in parallel and so that even when using the
`get` class method, we can still overrid this location if needed.